requirements.txt
matplotlib==3.3.4
numpy==1.20.1
opencv_python==4.5.3.56
Pillow==9.4.0
PyQt5==5.15.7
scikit_image==0.18.1
seaborn==0.11.1
